debug PID parameter.

+++ Hardware/TIMER/timer.h
@@ -6,7 +6,7 @@
 
 void Timer3_Init(u16 arr,u16 psc);
 
-extern double x_TargetAngle,y_TargetAngle;		
+extern float x_TargetAngle,y_TargetAngle;		
 #endif
